patch: Expose callback on selection (#57)
* patch: Expose callback on selection

Let client handle a callback to decide how to handle cytoscape
selections. Examples can be to show data for a node or an edge when it
is selected. We want to use rdf-terms outside the cytoscape component so
nodes are RdfIndividuals and edges are RdfTriples. Both are given as
argument to callback function in an RdfSelection class.

Exporting Details-view that accepts sparqlResult

* Remove rdf-store and override lodash.set to fix snyk

Cytoscape have a dependency to a vulnerable version of lodash.set. Using
override to fix the problem. See https://www.stefanjudis.com/today-i-learned/how-to-override-your-dependencys-dependencies/