fix: undefined values in rdftriple
Undefined values in rdfTriple in selection callback. Added
cytoscapeDataKeys to make code more robust against these errors in the
future. We might want to add even stronger typing to this in the future,
but it is nontrivial as cytoscape is js library that already has
semi-useful typescript library with ElementDefinition that contains
data-property with source, target and id members. The problem is that
the rest of the data object is untyped so we can use it however we want.
Someone with good typescript skills can propbably fix this in way that
let us keep cytoscapes typing and flexibility, but let us use types to
protect us from certain bugs like this.