chore: refactor code to transform rdf to cytoscape
Some predicates should not only add a simple datapoint to the subject
node corresponding to object value. New predicates dealing with svg,
connectors and the relations between them should be able to create
several new nodes and edges. Found a useful abstraction that let each
predicate be responsable for creating a list of ElementDefinition. If 2
predicates have the same subject, we will merge elementdefinition in a
small step at the end of the rdfTriples2Elements funciton.