feat: support for updates (#120)
Add functionality for updating graph visualization with the use of rdf triples. Since the triples we add might be dependent on each other and dependent on triples already present in the graph, we first swipe through incoming data and produce necessary data. In the next swipe, postprocessing, we create new elements / update existing to produce the final result. Most code in this branch is in postprocessing for handling rotation and svg triples that affect both parent image and it's connectors.

A lot of code is added to src/components/sparqlGraph/StoryWrapper.tsx. This is only to demo / test the possibilities we have when updating the graph this way