chore: Use n3 types for rdf triples
We need to seperate literals from iri object (will be done in seperate
PR). Found this to be a good opertunity to remove custome rdfTriple
which is simply "subject - predicate - object" where all 3 values are
string with the more complex types from n3-library. This library
includes types for bnodes, literals, named nodes and even varaible nodes
which may be useful in the future.