chore: Use n3 types for rdf triples (#133)
* chore: Use n3 types for rdf triples

We need to seperate literals from iri object (will be done in seperate
PR). Found this to be a good opertunity to remove custome rdfTriple
which is simply "subject - predicate - object" where all 3 values are
string with the more complex types from n3-library. This library
includes types for bnodes, literals, named nodes and even varaible nodes
which may be useful in the future.

* Fix connectors when converting to svg nodes

Had a bug where old edges would not hit correct connectors when
converting from plain node to new fancy node with connectors. Bug was
limited to the demo code in storywrapper.

* Use data factory

* Handle rotation without svgId