Make sure selection callback propegates
We need a selection callback when elements are selected. This callback
should be defined in "Props"-object required by all our graph components
(cytoscape force3d, gojs, ...), but it is not part of the state of these
components. Therfore we now use GraphProps as properties for these react
componants and keep GraphStateProps for other purposes. GraphProps is
just GraphStateProps with the additional callback.

Since our graph react components is created inside a react HOC we need
to forward the callback from registration inside the GraphComponent
outwards to the actual react component client is using.