feat: add support for visualization alternatives (#169)
* Extracted state management to separate components and isolated transformations

* CyGraph component and story added

* Update patchGraph

Basic svg example now works with hasSvg, hasConnector and
hasConnectorSuffix.

Updated test to verify

Future: Need to handle all permutations of existing and new predicates with both
add and remove

* Add suggestion to selection callback

* refactored patchGraph

* Cleanup patchGraph

* SVG node position works

* refactored patchGraph.test.tsx

* Make sure selection callback propegates

We need a selection callback when elements are selected. This callback
should be defined in "Props"-object required by all our graph components
(cytoscape force3d, gojs, ...), but it is not part of the state of these
components. Therfore we now use GraphProps as properties for these react
componants and keep GraphStateProps for other purposes. GraphProps is
just GraphStateProps with the additional callback.

Since our graph react components is created inside a react HOC we need
to forward the callback from registration inside the GraphComponent
outwards to the actual react component client is using.

* Add rotation to cy example

* Delete old sparqlGraph code


Co-authored-by: Eirik Alme Norstrand <einor@equinor.com>
Co-authored-by: dzianismakeichyk <dzianis.makeichyk@gmail.com>