feat: add gojs support (#164)
gojs specific code:

* Add support for gojs. This dependency is only a dev dependency and if clients want to use this, the library must be installed separately and appropriate license chosen as explained in readme.md. We currently do not provide a way for client to inject their license to remove watermark, but will fix this in the future. 

* sorted out working top-down DAG-layout in goGraph (layered, with edges entering on top and leaving below nodes, proper spacing between connecting points)

General code:

* Make sure selection callback propegates

We need a selection callback when elements are selected. This callback
should be defined in "Props"-object required by all our graph components
(cytoscape force3d, gojs, ...), but it is not part of the state of these
components. Therefore we now use GraphProps as properties for these react
components and keep GraphStateProps for other purposes. GraphProps is
just GraphStateProps with the additional callback.

Since our graph react components is created inside a react HOC we need
to forward the callback from registration inside the GraphComponent
outwards to the actual react component client is using.

* Future: Need to handle all permutations of existing and new predicates with both
add and remove

* Dark mode

* Moved selection-logic out of gojs, refactored to support propertyAssertion on edges, altered edge/connector logic in patchGraph to only supply front-end with a connector-embellished edge from node to node, retaining the connector-connector edge only in the propertygraph (probably too complicated, kind-of regret the connector object concept now)

Co-authored-by: Lorentz F. Barstad <29151133+lorentzfb@users.noreply.github.com>
Co-authored-by: Eirik Alme Norstrand <einor@equinor.com>
Co-authored-by: Martin Tofteberg Ulvesæter <48928133+toftalisk@users.noreply.github.com>