Add feat predicate props becomes edge props
We use metadata to store properties related to types and predicates.
These are not directly graph properties. Still, we used to yield them
and the gui implementations was responsable for interpreting it and
change properties to all elements with the corresponding types or all edges
that used the corresponding predicate. This responsability should be
moved to core and in this commit we solve the issue for predicates
(types are still remaining work)

An example is if a predicate is used twice resulting in two edges in the
corresponding graph. A new triple adding properties to the predicate
(using the predicate as subject) will result in 2 properties yielded to
the gui implementation. This brings us one step closer to the dream of
our gui implementation only implementing an interface with simple
methods like:

AddNode(<id>)
AddEdge(<id>)
AddProperty(<node- or edgeid>, <key>, <value>)
RemoveNode(<id>)
...

Added a new test strategy in src/core/tests/patchGraph/testUtils.ts. We
use simplified assertions to match against the actual assertions yielded
from patchGraph to check that we reseive the edge-, node- or
property-assertions in the right order. We do not inspect more content
opening some possibilites for bugs, but these tests with this
abstraction has proven to be useful.

Modified cygraph storywrapper to further tests the new functionality.
Has not yet been tested in the other GUI implementation other than not
breaking existing stories.