fix: remove direction when property is deleted
Problem when removing connector. Property removal hasRelativePosition
was correctly propegated, but direction not. This was because of the
check

```
if (prev)
```

since direction can have a legal value `0` which is falsy in javascript
this must be changed to

```
if (prev !== undefined)
```

Had a similar problem when adding properties, but this was fixed
earlier.

Also did minor refactoring (clear unused properties in patchGraph)