feat: Add common UI patcher and GoGraph UI patch handler (#217)
This PR adds an alternative way of handling communication of graph patches to the UI. This approach simplifies the UI implementations and centralise common UI logic.

Two main files have been added:

- src/core/ui/applyPatch.ts
This file lives in the "core" and can be used of all UI implementations (goGraph, cyGraph, etc.). This common function is responsible for invoking the UI (via an IUiPatchHandler) when an assertion is processed.

- src/goGraph/uiPatchHandler.ts
An implementation of the IUiPatchHandler which is a required argument on the UiNegotiator. A patch handler should only be required to implement the IUiPatchHandler for everything to work (synchronisation is handled by patchGraph).

Some other changes:

- Symbols for GoGraph use the 'geometryString' on a go.Shape instead of setting an SVG string as 'source' on go.Picture (Only two symbols have been converted: Compressor_1 and Valve_Gate_O). CyGraph not affected.
- Go node symbol template tuned to position connectors perfectly 🎉
- patchGraph prop assertions changed: relativePosition => connectorRelativePositon, direction => connectorDirection
- A symbolProvider can be injected as a prop on a UI implementation.


Problem when removing connector. Property removal hasRelativePosition
was correctly propegated, but direction not. This was because of the
check

```
if (prev)
```

since direction can have a legal value `0` which is falsy in javascript
this must be changed to

```
if (prev !== undefined)
```

Co-authored-by: toftalisk <martin.tofteberg@gmail.com>