feat: JSON Symbol library and symbol node styling (#229)
Initially, this PR was about improving styling of connector symbol nodes in goGraph. During this work, progress was made on the spec for the "smart" connector symbol SVG, and new symbols on this format was available to test. This led to more work especially on the symbol api implementation.

Changes

- New symbol library setup: Use a single json file as the symbol source (symbol-library.json). This file can be generated by using the Bo.Cli which batch processes SVGs on the new format and spits out a json file with a dictionary structure of symbols. Previously, we had one .ts file for each symbol. The new solution is more future proof in the sense that json is more likely to be the symbol exchange format.
- Old symbol library system removed.
- Added a storybook story for the symbol api to show all available symbols using RdfGoGraph (StoryWrapper for GoGraph reused). In this story, all symbols in symbol-library.json are converted to turtle at runtime which at the end becomes quads that are dispatched as a RdfActions. This gives us a useful overview of all symbols, and do visual/functional tests of new symbols with minimal effort.
- Symbol node template updated to handle new symbol format
- Added color and hover effects to the symbol node template
- CyGraph updated to handle new symbol api. 90% success rate (see further work below)
- Tests in core updated to handle new symbol api

Further work

Connector Symbol rotation does not work properly in CyGraph. The issue was introduced by the fact that the new symbol API can return symbols where width != height (we only used square shaped symbols in Cy and did not resolve this issue). Currently, width and height does not change when the symbol is rotated, and the symbol is clipped. The easiest fix is to allow only angles of n x 90 and then just flip width and height on rotation:
// Original w and h (at 0deg rotation)
const w = 200;
const h = 400;
// Width and height adjusted for 90deg rotation
const angle = 90;
const flip = (angle / 90) % 2 === 1;
const w_adj = flip ? h : w; // 400
const h_adj = flip ? w : h; // 200