feat: Expose gojs diagram (#255)
This PR exposes the go.Diagram object using a prop (diagramInitializer). The end user can instatiate a go.Diagram and pass it as a prop to the RdfGoGraph component (optional):

<RdfGoGraph
  {...state}
  diagramInitializer={() => diagramRef.current}
  containerStyle={diagramStyle}
  symbolProvider={symbolProviderJson}
  selectionEffect={handleSelection}
/>

The rdf-graph goGraph library is now only responsible for applying patches using the GoJsPatchHandler.

The advantage of this design is that we don't have to re-expose all kinds of go related functionality through RdfGoGraph. The user can take full advantage of GoJS. The responsibility of this library is to transform RDF to UI commands and invoke them, not deal with how a user should configure GoJS.

The consequence of this is that the storybook stories in rdf-graph will be quite simple, only showcasing the features of src/core. Thus, the ESD diagram story has been removed because this is not a core feature (involves special GoJS layout features, node templates etc.). For now, all Go templates related to Equinor work will be moved to https://github.com/equinor/spine-shutdown-hierarchies.