fix: bug when removing predicate data (#337)
If a node is completly detached from the rest of the graph, i.e. no
properties or edges to or from the node, it should be removed. The
detection of this state was buggy as we assumed the subject would exist
as node in the graph state since something has just been removed from
the subject. Either a property or a related edge. Forgot to think of the
case where the property was removed from as predicate node. In this case
the subject (referencing the predicate node) would most likely (always?)
be non existing in node store.

Added test and fixed bug