fix(go): port mapping bugfix in gojs impl (#367)
Attempt to fix bug in gojs where edges was connecting to the same port even though they was mapped to different ports. Seems to work.

Now, we ignore changing** existing links/edges that go to or from a node X when a "connectorName" prop on a connector node comes after an edge to/from X has been added. 

** Before, we used to set 'toPort' and/or 'fromPort' on links to/from X.



We cannot do this because there might be multiple links to/from the same node to/from different ports, and the deleted code in 'addConnectorNodeProp(...)' func will overwrite existing portIds for links to/from the node. This means that we relay on that the connectorNames that is available when the edge is added
```
function addEdge(diagram: go.Diagram, state: RdfGoGraphState, edge: GraphEdge) {
	// Check if 'from' and/or 'to' is a symbol connector
	const sourceConnector = state.connectorNodes[edge.sourceId];
	const targetConnector = state.connectorNodes[edge.targetId];
```
`sourceConnector` and `targetConnector` should have the portIds already.